+++ pkg/synthetics/secure_credentials.go
@@ -12,11 +12,15 @@ type SecureCredential struct {
 }
 
 // GetSecureCredentials is used to retrieve all secure credentials from your New Relic account.
+
+// Deprecated: Use entities.GetEntitySearch instead.
 func (s *Synthetics) GetSecureCredentials() ([]*SecureCredential, error) {
 	return s.GetSecureCredentialsWithContext(context.Background())
 }
 
 // GetSecureCredentialsWithContext is used to retrieve all secure credentials from your New Relic account.
+
+// Deprecated: Use entities.GetEntitySearchWithContext instead.
 func (s *Synthetics) GetSecureCredentialsWithContext(ctx context.Context) ([]*SecureCredential, error) {
 	resp := getSecureCredentialsResponse{}
 
@@ -29,11 +33,15 @@ func (s *Synthetics) GetSecureCredentialsWithContext(ctx context.Context) ([]*Se
 }
 
 // GetSecureCredential is used to retrieve a specific secure credential from your New Relic account.
+
+// Deprecated: Use entities.GetEntitySearch instead.
 func (s *Synthetics) GetSecureCredential(key string) (*SecureCredential, error) {
 	return s.GetSecureCredentialWithContext(context.Background(), key)
 }
 
 // GetSecureCredentialWithContext is used to retrieve a specific secure credential from your New Relic account.
+
+// Deprecated: Use entities.GetEntitySearchWithContext instead.
 func (s *Synthetics) GetSecureCredentialWithContext(ctx context.Context, key string) (*SecureCredential, error) {
 	var sc SecureCredential
 
@@ -46,11 +54,15 @@ func (s *Synthetics) GetSecureCredentialWithContext(ctx context.Context, key str
 }
 
 // AddSecureCredential is used to add a secure credential to your New Relic account.
+
+// Deprecated: Use synthetics.SyntheticsCreateSecureCredential instead.
 func (s *Synthetics) AddSecureCredential(key, value, description string) (*SecureCredential, error) {
 	return s.AddSecureCredentialWithContext(context.Background(), key, value, description)
 }
 
 // AddSecureCredentialWithContext is used to add a secure credential to your New Relic account.
+
+// Deprecated: Use synthetics.SyntheticsCreateSecureCredentialWithContext instead.
 func (s *Synthetics) AddSecureCredentialWithContext(ctx context.Context, key, value, description string) (*SecureCredential, error) {
 	sc := &SecureCredential{
 		Key:         key,
@@ -67,11 +79,15 @@ func (s *Synthetics) AddSecureCredentialWithContext(ctx context.Context, key, va
 }
 
 // UpdateSecureCredential is used to update a secure credential in your New Relic account.
+
+// Deprecated: Use synthetics.SyntheticsUpdateSecureCredential instead
 func (s *Synthetics) UpdateSecureCredential(key, value, description string) (*SecureCredential, error) {
 	return s.UpdateSecureCredentialWithContext(context.Background(), key, value, description)
 }
 
 // UpdateSecureCredentialWithContext is used to update a secure credential in your New Relic account.
+
+// Deprecated: Use synthetics.SyntheticsUpdateSecureCredentialWithContext instead
 func (s *Synthetics) UpdateSecureCredentialWithContext(ctx context.Context, key, value, description string) (*SecureCredential, error) {
 	sc := &SecureCredential{
 		Key:         key,
@@ -89,11 +105,15 @@ func (s *Synthetics) UpdateSecureCredentialWithContext(ctx context.Context, key,
 }
 
 // DeleteSecureCredential deletes a secure credential from your New Relic account.
+
+// Deprecated: Use synthetics.SyntheticsDeleteSecureCredential instead
 func (s *Synthetics) DeleteSecureCredential(key string) error {
 	return s.DeleteSecureCredentialWithContext(context.Background(), key)
 }
 
 // DeleteSecureCredentialWithContext deletes a secure credential from your New Relic account.
+
+// Deprecated: Use synthetics.SyntheticsDeleteSecureCredentialWithContext instead
 func (s *Synthetics) DeleteSecureCredentialWithContext(ctx context.Context, key string) error {
 	_, err := s.client.DeleteWithContext(ctx, s.config.Region().SyntheticsURL("/v1/secure-credentials", key), nil, nil)
 	if err != nil {